Re-organise
The start of an interface emerges.
