+++ sql/server/pg/src/test/kotlin/RepliTest.kt
@@ -1,6 +1,12 @@
 package io.skiplabs.skdb.pg
 
-@kotlin.ExperimentalStdlibApi
 fun main() {
-  connect("jdbc:postgresql://pg:5432/imdb")
+  val loop =
+      PgEventLoop(
+          url = "jdbc:postgresql://pg:5432/imdb",
+          user = "postgres",
+          password = "password",
+          slot = "skdb_test",
+          publication = "skdb_pub")
+  loop.start()
 }
